float to double, finished force and torque

+++ lib/Entity.cpp
@@ -5,7 +5,7 @@ Vector2 Entity::getSupport(Vector2 d) {
 
     auto current = vertices.begin();
     Vector2 support = *current;
-    float maxDot = support ^ d;
+    double maxDot = support ^ d;
     for (; current != vertices.end(); ++current) {
         if ( ((*current) ^ d) > maxDot) {
             support = *current;
